Rename localStorage service
