+++ client/src/components/Task/TaskDialog/CreateTaskDialog.vue
@@ -15,7 +15,7 @@
 
 <script setup>
 import Button from 'primevue/Button';
-import { localStorageAccount } from '../../service/localStorageAccount';
+import { localStorageAccount } from '../../service/localStorage';
 import { PrimeIcons } from 'primevue/api';
 import { ref } from 'vue';
 import taskApi from '@/src/api/tasks';
